refactor(build): give the bootstrap helper its own module
`libkrun_builder.rs` sat at 1491 of the 1500 production lines
`check-file-size` allows, so the preceding commit did not fit beside it.
The helper is not libkrun's anyway — `ensure_builder_vm_image` reaches it
on every backend — so acquisition, the payload-declaration seam, the
re-exec, the toolchain preflight and their tests move to
`mvm_build::builder_vm_bootstrap`.

Pure move: `builder_vm_source_checkout_root` becomes `pub(crate)`, the
env-var name and the resolver stay reachable for the two
`ensure_builder_vm_image` tests that still drive them, and the three
`mvm-cli` call sites follow the new path. `check-single-home`'s exemption
for the toolchain module follows it to `mvm-build`, same reason as before:
zig and rustup discovery lives in the real $HOME.
