+++ crates/mvm-build/src/lib.rs
@@ -89,6 +89,10 @@ pub mod template_reuse;
 /// Persistent ext4 image materialization for user-attached block volumes.
 pub mod volume_image;
 
+/// Acquiring and running the builder-VM bootstrap helper (gated by
+/// `builder-vm`). See module-level docs.
+#[cfg(feature = "builder-vm")]
+pub mod builder_vm_bootstrap;
 /// libkrun-backed builder VM (gated by `builder-vm`). See module-level
 /// docs.
 #[cfg(feature = "builder-vm")]
