+++ crates/mvm-cli/src/commands/mod.rs
@@ -555,7 +555,7 @@ fn register_builder_session_starter() {}
 /// helper, and the compile is minutes spent reproducing what is already loaded.
 #[cfg(feature = "builder-vm")]
 fn declare_embedded_host_binaries() {
-    mvm_build::libkrun_builder::declare_current_exe_carries_host_binaries(
+    mvm_build::builder_vm_bootstrap::declare_current_exe_carries_host_binaries(
         !crate::host_binaries::embedded::EMBEDDED.is_empty(),
     );
 }
